ESP32 code for the smart chair project
A smart chair IoT project. The smart chair detects when someone is sitting on it, as well as detecting if a person is at their desk. Measures length of sitting as well as how long a person has been at their desk. Gives reminders to stand/get up after a certain amount of continuous sitting/continuous desk activity. Also measures air quality using the CCS811 sensor, so that proper ventilation and air quality can be maintained and monitored in a work space.